Flatten ut the folder stucture to be components, pages, hooks, services, and utils
