+++ frontend/dashboard/app/App.tsx
@@ -5,7 +5,7 @@ import type { IHeaderContext } from 'app-shared/navigation/main-header/Header';
 import type { SelectedContext } from '../resources/fetchDashboardResources/dashboardSlice';
 import { AltinnSpinner } from 'app-shared/components';
 import { Button } from '@digdir/design-system-react';
-import { CenterContainer } from '../common/components/CenterContainer';
+import { CenterContainer } from '../components/CenterContainer';
 import { CreateService } from '../pages/CreateService';
 import { Dashboard } from '../pages/Dashboard';
 import { DashboardActions } from '../resources/fetchDashboardResources/dashboardSlice';
@@ -14,9 +14,10 @@ import { Route, Routes } from 'react-router-dom';
 import { fetchLanguage } from '../resources/fetchLanguage/languageSlice';
 import { getLanguageFromKey } from 'app-shared/utils/language';
 import { post } from 'app-shared/utils/networking';
-import { useAppDispatch, useAppSelector } from '../common/hooks';
+import { useAppDispatch } from '../hooks/useAppDispatch';
+import { useAppSelector } from '../hooks/useAppSelector';
 import { useGetOrganizationsQuery } from '../services/organizationApi';
-import { userHasAccessToSelectedContext } from '../common/utils';
+import { userHasAccessToSelectedContext } from '../utils/userUtils';
 import AppHeader, {
   HeaderContext,
   SelectedContextType
