+++ frontend/dashboard/components/CenterContainer/index.ts
@@ -1,0 +1,1 @@
+export * from './CenterContainer';
