+++ frontend/dashboard/components/Footer/Footer.module.css
