+++ frontend/dashboard/components/MakeCopyModal/index.ts
@@ -1,0 +1,1 @@
+export * from './MakeCopyModal';
