+++ frontend/dashboard/components/OrgRepoList/OrgReposList.tsx
@@ -1,10 +1,11 @@
 import React, { useState } from 'react';
 import type { GridSortModel } from '@mui/x-data-grid';
 import { DashboardActions } from '../../resources/fetchDashboardResources/dashboardSlice';
-import { RepoList } from '../../common/components/RepoList';
+import { RepoList } from '../../components/RepoList';
 import { getReposLabel } from '../../utils/repoUtils';
 import { getUidFilter } from '../../utils/filterUtils';
-import { useAppDispatch, useAppSelector } from '../../common/hooks';
+import { useAppDispatch } from '../../hooks/useAppDispatch';
+import { useAppSelector } from '../../hooks/useAppSelector';
 import { useAugmentReposWithStarred } from '../../hooks/useAugmentReposWithStarred';
 import { useGetOrganizationsQuery } from '../../services/organizationApi';
 import { useGetSearchQuery } from '../../services/repoApi';
