+++ frontend/dashboard/components/ResourceItem/ResourceItem.module.css
