+++ frontend/dashboard/components/ResourceItem/index.ts
@@ -1,0 +1,1 @@
+export * from './ResourceItem';
