+++ frontend/dashboard/components/Resources/Resources.module.css
