+++ frontend/dashboard/hooks/useAppDispatch/index.ts
@@ -1,0 +1,1 @@
+export * from './useAppDispatch';
