+++ frontend/dashboard/hooks/useAppSelector/index.ts
@@ -1,0 +1,1 @@
+export * from './useAppSelector';
