+++ frontend/dashboard/hooks/useAppSelector/useAppSelector.ts
@@ -1,0 +1,6 @@
+import { useSelector } from 'react-redux';
+import type { TypedUseSelectorHook } from 'react-redux';
+import type { RootState } from '../../app/store';
+
+// Use throughout your app instead of plain `useSelector`
+export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
