+++ frontend/dashboard/index.tsx
@@ -3,10 +3,10 @@ import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
 import throttle from 'lodash-es/throttle';
+import { DASHBOARD_BASENAME } from 'app-shared/constants';
 import { App } from './app/App';
 import { run } from './app/rootSaga';
-import { loadFromLocalStorage, saveToLocalStorage } from './common/utils/localStorage';
-import { DASHBOARD_BASENAME } from 'app-shared/constants';
+import { loadFromLocalStorage, saveToLocalStorage } from './utils/localStorageUtils';
 import { setupStore } from './app/store';
 
 const store = setupStore(loadFromLocalStorage());
