+++ frontend/dashboard/utils/urlUtils/index.ts
@@ -1,0 +1,1 @@
+export * from './urlUtils';
